timeout fixed and permissions added

+++ funcs.js
@@ -19,31 +19,47 @@ module.exports = {
             if (msg.content.toLowerCase().includes(word)) {
                 switch (word) {
                     case "ban":
-                        banUser(word, msg, client);
+                        if(hasPermission(msg.member,"BAN_MEMBERS")){
+                            banUser(word, msg, client);
+                        }
                         break;
 
                     case "kick":
-                        kickUser(word, msg, client);
+                        if(hasPermission(msg.member,"KICK_MEMBERS")){
+                            kickUser(word, msg, client);
+                        }
                         break;
 
                     case "timeout":
-                        timeoutUser(word, msg, client);
+                    
+                        if(hasPermission(msg.member,"MANAGE_MESSAGES")){
+                            timeoutUser(word, msg, client);
+                        }
                         break;
 
                     case "enable" || "activate":
-                        featureToggle(msg, true);
+                        if(hasPermission(msg.member,"ADMINISTRATOR")){
+                            featureToggle(msg, true);
+                        }
                         break;
 
                     case "disable" || "deactivate":
-                        featureToggle(msg, false);
+                        
+                        if(hasPermission(msg.member,"ADMINISTRATOR")){
+                            featureToggle(msg, false);
+                        }
                         break;
 
                     case "turn":
-                        featureToggle(msg, null);
+                        if(hasPermission(msg.member,"ADMINISTRATOR")){
+                            featureToggle(msg, null);
+                        }
                         break;
 
                     case "purge":
-                        purgeChannel(msg, client)
+                        if(hasPermission(msg.member,"MANAGE_MESSAGES")){
+                            purgeChannel(msg, client)
+                        }
                         break;
                 }
             }
@@ -271,11 +287,13 @@ function timeoutUser(word, msg, client) {
 
         ids.forEach(function (id) {
             if (id != "592783579998584868") {
-                end = calculateTimeoutEnd(id, timer);
-                if (end != 0) {
-                    var member = msg.guild.members.get(id);
-                    timeouts.push(member.displayName);
-                    member.send("You have been timed out from **" + client.guilds.get(msg.guild.id).name + "**, you will be able to chat again in **" + end + "**");
+                if(isTimedout(id) == false){
+                    end = calculateTimeoutEnd(id, timer);
+                    if (end != 0) {
+                        var member = msg.guild.members.get(id);
+                        timeouts.push(member.displayName);
+                        member.send("You have been timed out from **" + client.guilds.get(msg.guild.id).name + "**, you will be able to chat again in **" + end + "**");
+                    }
                 }
 
             }
@@ -321,7 +339,7 @@ function calculateTimeoutEnd(did, times) {
         });
     }
 
-    if (m != 0) {
+    if (s != 0) {
         times.seconds.forEach(function (seconds) {
             suc = suc + parseInt(seconds.replace(/\D/g, ''));
         });
@@ -453,7 +471,6 @@ function featureToggle(msg, toggleType) {
     }
 }
 
-
 function getTranslatedText(msg, callback) {
     var translated;
 
@@ -530,12 +547,13 @@ function featureToggle(msg, toggleType) {
 
 function purgeChannel(msg, client) {
     if (msg.mentions.channels.size > 0) {
-        var channelId = msg.mentions.channels.array()[0].id;
-        client.guilds.get(msg.guild.id).channels.get(channelId).fetchMessages().then(messages => messages.array().forEach(message => {
-            message.delete();
+        msg.mentions.channels.forEach((function(channel){
+            var channelId = channel.id;
+            client.guilds.get(msg.guild.id).channels.get(channelId).fetchMessages().then(messages => messages.array().forEach(message => {
+                message.delete();
+            }));
         }));
-    }
-    else if (msg.mentions.members.size > 0) {
+    }else if (msg.mentions.members.size > 0) {
         if (msg.mentions.members.array()[0].user.id == client.id) {
             if (msg.mentions.members.array()[1].exists()) {
                 var userid = msg.mentions.members.array()[1].user.id;
@@ -545,8 +563,7 @@ function purgeChannel(msg, client) {
                     }
                 }));
             }
-        }
-        else {
+        }else {
             var userid = msg.mentions.members.array()[0].user.id;
             msg.channel.fetchMessages().then(messages => messages.array().forEach(message => {
                 if (message.author.id == userid) {
@@ -555,4 +572,22 @@ function purgeChannel(msg, client) {
             }));
         }
     }
+}
+
+function isTimedout(id){
+    var isTimedout = false;
+    var filedata = fs.readFileSync('./timeouts.json', { encoding: 'utf8' });
+    var timeouts = JSON.parse(filedata);
+    timeouts["active"].forEach(function (obj) {
+        if (obj.id == id) {
+            var timeEnd = new Date(obj.end);
+            var now = new Date();
+            isTimedout = true;
+        }
+    });
+    return isTimedout;
+}
+
+function hasPermission(member,permission){
+    return member.hasPermission(permission) ? true : false;
 }
